Use ActivityTestRule in ContactActivitySteps
