+++ app/src/androidTest/java/com/nervousfish/nervousfish/test/Instrumentation.java
@@ -1,9 +1,12 @@
 package com.nervousfish.nervousfish.test;
 
 import android.os.Bundle;
+import android.support.test.espresso.intent.Intents;
 import android.support.test.runner.MonitoringInstrumentation;
 
 import cucumber.api.android.CucumberInstrumentationCore;
+import cucumber.api.java.After;
+import cucumber.api.java.Before;
 
 public class Instrumentation extends MonitoringInstrumentation {
 
@@ -23,6 +26,16 @@ public class Instrumentation extends MonitoringInstrumentation {
 
         waitForIdleSync();
         mInstrumentationCore.start();
+    }
+
+    @Before
+    public void setUp() throws Exception {
+        Intents.init();
+    }
+
+    @After
+    public void tearDown() {
+        Intents.release();
     }
 
 }
